[CLEAN] remove main, replace println with debug
