+++ src/lib.rs
@@ -16,3 +16,65 @@
  */
 
 pub mod minio;
+
+#[cfg(test)]
+mod tests {
+    use futures::{future::Future, stream::Stream};
+    use hyper::rt;
+    use log::debug;
+
+    use minio::BucketInfo;
+
+    use super::*;
+
+    fn get_local_default_server() -> minio::Client {
+        match minio::Client::new("http://localhost:9000") {
+            Ok(mut c) => {
+                c.set_credentials(minio::Credentials::new("minio", "minio123"));
+                c
+            }
+            Err(_) => panic!("could not make local client"),
+        }
+    }
+
+    #[test]
+    fn test_lib_functions() {
+        rt::run(rt::lazy(|| {
+            let c = minio::Client::get_play_client();
+            let bucket = "aaaa";
+
+            c.put_object_req(bucket, "hhhhhhhhhh", vec![], "object content".as_bytes().to_vec())
+            .and_then(|g| {
+                debug!("object: {} {} {:?}", g.object_size, g.etag, g.content_type);
+                g.get_object_stream().concat2()
+            })
+            .map(|c| debug!("get obj res: {:?}", c))
+            .map_err(|c| debug!("err res: {:?}", c))
+            .map(|_| {})
+        }));
+
+        rt::run(rt::lazy(|| {
+            let c = minio::Client::get_play_client();
+            let bucket = "aaaa";
+
+            c.get_object_req(bucket, "hhhhhhhhhh", vec![])
+            .and_then(|g| {
+                debug!("object: {} {} {:?}", g.object_size, g.etag, g.content_type);
+                g.get_object_stream().concat2()
+            })
+            .map(|c| debug!("get obj res: {:?}", c))
+            .map_err(|c| debug!("err res: {:?}", c))
+            .map(|_| {})
+        }));
+
+        rt::run(rt::lazy(|| {
+            let c = minio::Client::get_play_client();
+            let bucket = "aaaa";
+
+            c.delete_bucket(bucket)
+            .map(|_| debug!("Deleted!"))
+            .map_err(|err| debug!("del err: {:?}", err))
+            .map(|_| {})
+        }));
+    }
+}
